resolve: Move `get_traits_in_module_containing_item` to Resolver
